Add explicit constructors to satisfy Javadoc (#118)
By testing the code base using Java 21 (the default version on the most
recent Debian), it was revealed that its Javadoc version generates many
warnings.  Most were due to classes "using a default constructor which
does not provide a comment"; this was addressed by adding an explicit
contructor with a trivial comment (perhaps a better one can be added,
for some of them).  A couple of few other missing comments were added.